requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
openai==1.6.1
anthropic==0.8.1

# Chainlit for UI
chainlit==1.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tiktoken==0.5.2
tenacity==8.2.3
httpx>=0.23.0,<0.25.0

# Development tools (optional)
black==23.12.1
isort==5.13.2
mypy==1.8.0

